+++ changelog.md
@@ -5,3 +5,4 @@
 - Added Turkish translation
 - Fixed custom name rendering
 - Fixed name showing on top of the plane
+- Don't trim plane names